Bump dependencies
This change updates the dependencies and fixes the test to run with the
updated code and dependencies, specifically this removes a dependency
on github.com/fluxcd/toolkit which was bringing in an old (insecure)
version of github.com/fluxcd/kustomize-controller.

The YAML changes were needed because we're now getting indented lists
and dictionaries.